Refactor Entities using Traits to reduce redundancy
