+++ module/Order/src/Order/Entity/Traits/CreatedDateTrait.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Order\Entity\Traits;
+
+use DateTime;
+use Doctrine\ORM\Mapping as ORM;
+
+trait CreatedDateTrait
+{
+    /**
+     * @ORM\Column(type="datetime")
+     **/
+    protected $created;
+
+    /**
+     * @return DateTime
+     */
+    public function getCreated()
+    {
+        return $this->created;
+    }
+
+    /**
+     * @ORM\PrePersist
+     */
+    public function setCreated()
+    {
+        $this->created = new DateTime();
+    }
+}
